requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
pydantic-settings
pillow
langchain
openai
langchain-openai
numpy
sqlalchemy
aiohttp
aiofiles
httpx
tiktoken
diffusers
transformers
accelerate
torch